+++ horoshilov.feodor/T2/Guard.h
